+++ src/functions.jl
@@ -146,3 +146,29 @@ basetype(rand(Complex{Float64})) === Complex
 basetype(t::DataType) = t.name.wrapper
 basetype(t::UnionAll) = basetype(t.body)
 basetype(::T) where T = basetype(T)
+
+"""
+    asdeg(x::Real)
+
+Convert the provided value assumed to be in radians to Unitful degrees.
+
+The [`stripdeg`](@ref) function performs the inverse operation.
+
+```julia
+asdeg(π) ≈ 180.0°
+```
+"""
+asdeg(x::Real) = rad2deg(x) * °
+
+"""
+    stripdeg(x::Deg)
+
+Strip the units from the provided `Deg` field and convert it to radians.
+
+The [`asdeg`](@ref) function performs the inverse operation.
+
+```julia
+stripdeg(180.0°) ≈ π
+```
+"""
+stripdeg(x::Deg) = x |> ustrip |> deg2rad